Do not auto generate .revision with specifically being configed
No longer clobber the .revision file with the latest timestamp on every dist.
There are instances like when PWA is enabled during development and you
want to force the cache to flush on every dist but this is unusual to use
on production or when using start.
It is preferred to have the git hash used via Beanstalk or a deploy script.

To enable this feature add "regenerate_revision":true to the package.json
of the project.

This way we can also selectively enable/disable this per environment via a
local-config.json.