Include postcss-focus-visible plugin in PostCSS config
`postcss-focus-visible` requires an external polyfill which is not
loaded by this library.

We don't include the polyfill, because Some sites will not use this
feature and others only load the JS app on particular pages which
creates a delta
between which CSS we can use in JS app vs CSS app, unless we catch this
and load the polyfill on every page.

After much consideration it was decided to include the PostCSS plugin in
both libraries and load the polyfill only as needed.

Once Safari starts supporting `:focus-visible` we can remove the
plugin, and any uses of the polyfill.